+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/triggers/TimerTriggerStrategy.kt
@@ -6,11 +6,14 @@ import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.FinalConfig
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.PostId
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.*
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.abstractions.Chooser
+import com.github.insanusmokrassar.AutoPostTelegramBot.mediumBroadcastCapacity
 import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.publishers.Publisher
+import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.scheduler.SchedulerPlugin
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.*
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.commands.Command
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.nearDateTime
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.sendToLogger
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.flow.collectWithErrors
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
 import com.github.insanusmokrassar.TelegramBotAPI.requests.ForwardMessage
 import com.github.insanusmokrassar.TelegramBotAPI.requests.send.SendMessage
@@ -21,6 +24,9 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Common
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.TextContent
 import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.executeUnsafe
 import kotlinx.coroutines.*
+import kotlinx.coroutines.channels.BroadcastChannel
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.flow.asFlow
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
 import org.joda.time.*
@@ -28,13 +34,13 @@ import kotlin.collections.component1
 import kotlin.collections.component2
 import kotlin.collections.set
 
-private val TimerTriggerStrategyScope = NewDefaultCoroutineScope(1)
+private val TimerTriggerStrategyScope = CoroutineScope(Dispatchers.Default)
 
 private val timerScheduleCommandRegex = Regex("((getA)|(a))utoPublication(s [\\d]+)?$")
 private val numberRegex: Regex = Regex("[\\d]+$")
 private class TimerScheduleCommand(
     private val sourceChatId: ChatId,
-    private val timesOfTriggering: List<CalculatedDateTime>,
+    private val timeChooserCallback: (after: DateTime) -> DateTime?,
     private val chooser: Chooser,
     private val executor: RequestsExecutor,
     private val postsMessagesTable: PostsMessagesTable,
@@ -48,7 +54,7 @@ private class TimerScheduleCommand(
         var nearDateTime = DateTime.now()
         val chosen = mutableMapOf<DateTime, List<PostId>>()
         for (i in 0 until count) {
-            nearDateTime = timesOfTriggering.nearDateTime(nearDateTime + 1000L) // plus one second
+            nearDateTime = timeChooserCallback(nearDateTime + 1000L) // plus one second
             chosen[nearDateTime] = chooser.triggerChoose(nearDateTime, chosen.values.flatten()).toList()
         }
         commonLogger.info("Was requested auto publishing count: $count; Was chosen to show: ${chosen.values.sumBy { it.size }}")
@@ -94,30 +100,69 @@ private fun getPlusTwentyFourHours() = System.currentTimeMillis() + twentyFourHo
 @Serializable
 class TimerTriggerStrategy (
     private val delay: Long? = null,
-    private val time: String = "00:00-00:00 01:00"
+    private val time: String = "00:00-00:00 01:00",
+    private val substitutedByScheduler: Boolean = false
 ) : Plugin {
+    @Transient
+    private val publicationTimesPossiblyChangedBroadcastChannel = BroadcastChannel<DateTime>(
+        mediumBroadcastCapacity
+    )
+    @Transient
+    val triggerTimesAffectedFlow = publicationTimesPossiblyChangedBroadcastChannel.asFlow()
+
     @Transient
     private var lastTime = DateTime.now()
 
     private val timesOfTriggering: List<CalculatedDateTime> by lazy {
         time.parseDateTimes()
     }
+    fun getNextTime(after: DateTime = DateTime.now()): DateTime? {
+        val near = timesOfTriggering.nearDateTime(after) ?: return null
+        schedulerPluginToCheckCollision ?.let {
+            val thereIsScheduled = it.timerSchedulesTable.registeredPostsTimes(
+                near - 1000L to near + 1000L
+            ).any { (_, dateTime) ->
+                dateTime == near
+            }
+            if (thereIsScheduled) {
+                return getNextTime(near)
+            }
+        }
+        return near
+    }
     val nextTriggerTime: DateTime?
         get() = delay ?.let {
             lastTime = lastTime.plus(it)
             lastTime
-        } ?: timesOfTriggering.nearDateTime()
+        } ?: getNextTime()
 
     @Transient
     private lateinit var timerScheduleCommand: TimerScheduleCommand
+    private var schedulerPluginToCheckCollision: SchedulerPlugin? = null
 
     override suspend fun onInit(executor: RequestsExecutor, baseConfig: FinalConfig, pluginManager: PluginManager) {
         val publisher: Publisher = pluginManager.findFirstPlugin() ?: return
         val chooser: Chooser = pluginManager.findFirstPlugin() ?: return
 
+        if (substitutedByScheduler) {
+            pluginManager.findFirstPlugin<SchedulerPlugin>() ?.let {
+                schedulerPluginToCheckCollision = it
+                TimerTriggerStrategyScope.launch {
+                    it.timerSchedulesTable.postTimeRegisteredFlow.collectWithErrors { (_, dateTime) ->
+                        publicationTimesPossiblyChangedBroadcastChannel.send(dateTime)
+                    }
+                }
+                TimerTriggerStrategyScope.launch {
+                    it.timerSchedulesTable.postTimeChangedFlow.collectWithErrors { (_, dateTime) ->
+                        publicationTimesPossiblyChangedBroadcastChannel.send(dateTime)
+                    }
+                }
+            }
+        }
+
         timerScheduleCommand = TimerScheduleCommand(
             baseConfig.sourceChatId,
-            timesOfTriggering,
+            ::getNextTime,
             chooser,
             executor,
             PostsMessagesTable,
@@ -125,8 +170,7 @@ class TimerTriggerStrategy (
         )
 
         TimerTriggerStrategyScope.launch {
-            while (isActive) {
-                delay(nextTriggerTime ?.millis ?.minus(System.currentTimeMillis()) ?: break)
+            suspend fun trigger() {
                 launch {
                     try {
                         chooser.triggerChoose().forEach {
@@ -137,16 +181,36 @@ class TimerTriggerStrategy (
                     }
                 }
             }
+            triggerLoop@while (isActive) {
+                val currentNow = DateTime.now()
+                val nextTime = getNextTime(currentNow) ?: break
+                delay(nextTime.millis - currentNow.millis)
+                val nextAfterDelay = getNextTime(currentNow)
+                when {
+                    nextAfterDelay == nextTime -> trigger()
+                    nextAfterDelay == null || nextTime < getNextTime(currentNow) -> continue@triggerLoop
+                    else -> {
+                        var currentNext = nextAfterDelay
+                        do {
+                            currentNext = getNextTime(currentNext ?: break)
+                            if (currentNext == nextTime) {
+                                trigger()
+                                continue@triggerLoop
+                            }
+                        } while (currentNext != null && currentNext < nextTime)
+                    }
+                }
+            }
         }
     }
 
     fun getTriggersInRange(from: DateTime = DateTime.now(), to: DateTime = DateTime(getPlusTwentyFourHours())): List<DateTime> {
         val times = mutableListOf<DateTime>()
-        var current = timesOfTriggering.nearDateTime(from) ?: return times
+        var current = getNextTime(from) ?: return times
         val toMillis = to.millis
         while (current.millis <= toMillis) {
             times.add(current)
-            current = timesOfTriggering.nearDateTime(current + 1000L) ?: return times
+            current = getNextTime(current + 1000L) ?: return times
         }
         return times
     }